Simple permission sanity checks for log dir and config.
It checks when the config file has permissions set for others and
removes them, if it's the case. For log directory it ensures that 0o755
is set, so the logforwarder can access them later (in concunction on
ACLs). It also checks if the fileowner is root, otherwise it should
match the effective uid.