Coalesce: Deal with missing EOEs
Timeout handling for multi-part events is added. Messages are flushed
out via emit_event() when that timeout is reached.

The Drop trait is implemented. Messages are flushed out when the
Coalesce struct is dropped, e.g. on program exit.

To avoid duplicate code, we move message handling associated with EOE
to emit_event(): This includes touching up messages and feeding the
shadow process table.

Close #16