Fix race condition in logfile rotation
Creating the file as seen by "tail --follow=name" or a log forwarder
was non-atomic: The peremissive ACL is set after the new file has
appeared in the directory. Therefore, the reader could fail to open
the newly creeated file due to wrong permissions.

A temporary file is now used to set up permissions and it is moved
into place atomically.