Analyse exec events for script execution
For execve events, assume that differences between the file referenced
by SYSCALL.exe and the first PATH entry indicate that a script is
being invoked through the "#!" mechanism.

If the PATH entry contains relative path, ".", "..", and double
slashes are normalized.

This information is used to add SYSCALL.SCRIPT and to add process
labels.

Caveats:

1. Inspecting SYSCALL.exe is racy.

2. SYSCALL.exe is looked at through the /proc/<pid>/root "lens" to try
to treat mount namespaces used by container runtimes containers
properly. Unfortunately, this does not seem to work for Podman if
fuse-overlayfs is used. (Tested with podman/3.0.1+dfsg1-3+deb11u1,
fuse-overlayfs/1.4.0-1 on Debian/bullseye.)